+++ src/test/unit/maths_unittest.cc
@@ -207,8 +207,8 @@ TEST(MathsUnittest, TestFastTrigonometrySinCos)
 {
     double sinError = 0;
     for (float x = -10 * M_PI; x < 10 * M_PI; x += M_PI / 300) {
-        double approxResult = sin_approx(x);
-        double libmResult = sinf(x);
+        float approxResult = sin_approx(x);
+        double libmResult = sin(x);
         sinError = MAX(sinError, fabs(approxResult - libmResult));
     }
     printf("sin_approx maximum absolute error = %e\n", sinError);
@@ -216,8 +216,8 @@ TEST(MathsUnittest, TestFastTrigonometrySinCos)
 
     double cosError = 0;
     for (float x = -10 * M_PI; x < 10 * M_PI; x += M_PI / 300) {
-        double approxResult = cos_approx(x);
-        double libmResult = cosf(x);
+        float approxResult = cos_approx(x);
+        double libmResult = cos(x);
         cosError = MAX(cosError, fabs(approxResult - libmResult));
     }
     printf("cos_approx maximum absolute error = %e\n", cosError);
@@ -229,8 +229,8 @@ TEST(MathsUnittest, TestFastTrigonometryATan2)
     double error = 0;
     for (float x = -1.0f; x < 1.0f; x += 0.01) {
         for (float y = -1.0f; x < 1.0f; x += 0.001) {
-            double approxResult = atan2_approx(y, x);
-            double libmResult = atan2f(y, x);
+            float approxResult = atan2_approx(y, x);
+            double libmResult = atan2(y, x);
             error = MAX(error, fabs(approxResult - libmResult));
         }
     }
@@ -242,7 +242,7 @@ TEST(MathsUnittest, TestFastTrigonometryACos)
 {
     double error = 0;
     for (float x = -1.0f; x < 1.0f; x += 0.001) {
-        double approxResult = acos_approx(x);
+        float approxResult = acos_approx(x);
         double libmResult = acos(x);
         error = MAX(error, fabs(approxResult - libmResult));
     }
